+++ Tracker/TrackerCategoryStore.swift
@@ -1,0 +1,27 @@
+import Foundation
+import CoreData
+
+/// Класс, работающий с категориями в БД
+final class TrackerCategoryStore {
+    
+    // MARK: - Метод, добавляющий категорию в БД
+    func addCategory(category: String, tracker: TrackerCoreData, context: NSManagedObjectContext) {
+        let request = NSFetchRequest<TrackerCategoryCoreData>(entityName: "TrackerCategoryCoreData")
+        request.returnsObjectsAsFaults = false
+        let trackerCategories = try! context.fetch(request)
+        if !trackerCategories.filter({$0.name == category}).isEmpty {
+            trackerCategories.forEach { trackerCategory in
+                if trackerCategory.name == category {
+                    trackerCategory.addToTrackers(tracker)
+                }
+            }
+        } else {
+            let newCategory = TrackerCategoryCoreData(context: context)
+            newCategory.name = category
+            newCategory.id = UUID()
+            newCategory.addToTrackers(tracker)
+        }
+        try! context.save()
+    }
+    
+}
